Combine ensureEnv.js and env.js.
