+++ src/commands/functions-env-set.ts
@@ -4,11 +4,10 @@ import * as fs from "fs";
 import * as path from "path";
 
 import { Command } from "../command";
-import { ensureEnvStore } from "../functions/ensureEnv";
 import { logger } from "../logger";
 import { promptOnce } from "../prompt";
 import { requirePermissions } from "../requirePermissions";
-import * as fenv from "../functions/env";
+import * as env from "../functions/env";
 import * as getProjectId from "../getProjectId";
 import * as utils from "../utils";
 
@@ -21,7 +20,7 @@ export default new Command("functions:env:set [values...]")
     "firebase.envstores.delete",
     "firebase.envstores.update",
   ])
-  .before(ensureEnvStore)
+  .before(env.ensureEnvStore)
   .action(async (args: string[], options: { file?: string; force: boolean }) => {
     if (!options.file && !args.length) {
       return utils.reject("Must supply at least one key/value pair, e.g. " + clc.bold("FOO=bar"));
@@ -31,11 +30,11 @@ export default new Command("functions:env:set [values...]")
       const buf = fs.readFileSync(path.resolve(options.file), "utf8");
       setEnvs = dotenv.parse(buf.toString().trim());
     } else {
-      setEnvs = fenv.parseKvArgs(args);
+      setEnvs = env.parseKvArgs(args);
     }
 
     const projectId = getProjectId(options);
-    const curEnvs = await fenv.getEnvs(projectId);
+    const curEnvs = await env.getEnvs(projectId);
 
     if (curEnvs && Object.keys(curEnvs).length) {
       const confirm = await promptOnce(
@@ -45,9 +44,9 @@ export default new Command("functions:env:set [values...]")
           default: false,
           message:
             "You about to replace current set of environment variables:\n" +
-            `\n${fenv.formatEnv(curEnvs)}\n\n` +
+            `\n${env.formatEnv(curEnvs)}\n\n` +
             "with the following environment variables:\n" +
-            `\n${fenv.formatEnv(setEnvs)}\n\n` +
+            `\n${env.formatEnv(setEnvs)}\n\n` +
             "Are you sure you want to do this?",
         },
         options
@@ -57,7 +56,7 @@ export default new Command("functions:env:set [values...]")
       }
     }
 
-    const envs = await fenv.setEnvs(projectId, setEnvs);
-    logger.info(fenv.formatEnv(envs));
+    const envs = await env.setEnvs(projectId, setEnvs);
+    logger.info(env.formatEnv(envs));
     return envs;
   });
